Move the migration logic to the "Migration" namespace
